simplify benchmark code
