--- bench/benches/clustered/dns_traffic/mod.rs
@@ -1,2 +1,0 @@
-mod data;
-pub mod sequential;
