--- bench/benches/clustered/world_cities/mod.rs
@@ -1,3 +1,0 @@
-pub mod bulkload;
-mod data;
-pub mod sequential;
